:sparkles: Feat: 스케줄 생성 api 개발중

+++ BACKEND/watchify/src/main/java/com/watchify/watchify/dto/request/ScheduleCreateRequestDTO.java
@@ -12,7 +12,7 @@ import java.util.Map;
 public class ScheduleCreateRequestDTO {
 
     private LocalDate startDate;
-    private List<Integer> contents;
+    private List<Long> contents;
     private List<Integer> patterns;
     private List<String> ott;
